+++ tests/test_flask_restx_mongo_controller_date.py
@@ -1,5 +1,7 @@
+import datetime
+
 import flask
-import flask_restplus
+import flask_restx
 import pytest
 
 import layabase
@@ -7,14 +9,13 @@ import layabase.mongo
 
 
 @pytest.fixture
-def controller():
+def controller() -> layabase.CRUDController:
     class TestCollection:
         __collection_name__ = "test"
 
-        float_key = layabase.mongo.Column(float, is_primary_key=True)
-        float_with_default = layabase.mongo.Column(float, default_value=34)
-        dict_field = layabase.mongo.Column(dict, is_required=True)
-        list_field = layabase.mongo.Column(list, is_required=True)
+        key = layabase.mongo.Column(str, is_primary_key=True)
+        date_str = layabase.mongo.Column(datetime.date)
+        datetime_str = layabase.mongo.Column(datetime.datetime)
 
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])
@@ -22,30 +23,30 @@ def controller():
 
 
 @pytest.fixture
-def app(controller):
+def app(controller: layabase.CRUDController):
     application = flask.Flask(__name__)
     application.testing = True
-    api = flask_restplus.Api(application)
+    api = flask_restx.Api(application)
     namespace = api.namespace("Test", path="/")
 
-    controller.namespace(namespace)
+    controller.flask_restx.init_models(namespace)
 
     @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
+    class TestResource(flask_restx.Resource):
+        @namespace.expect(controller.flask_restx.query_get_parser)
+        @namespace.marshal_with(controller.flask_restx.get_response_model)
         def get(self):
             return []
 
-        @namespace.expect(controller.json_post_model)
+        @namespace.expect(controller.flask_restx.json_post_model)
         def post(self):
             return []
 
-        @namespace.expect(controller.json_put_model)
+        @namespace.expect(controller.flask_restx.json_put_model)
         def put(self):
             return []
 
-        @namespace.expect(controller.query_delete_parser)
+        @namespace.expect(controller.flask_restx.query_delete_parser)
         def delete(self):
             return []
 
@@ -94,32 +95,25 @@ def test_open_api_definition(client):
                     "operationId": "delete_test_resource",
                     "parameters": [
                         {
-                            "name": "float_key",
+                            "name": "key",
                             "in": "query",
                             "type": "array",
-                            "items": {"type": "number"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "float_with_default",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "number"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "dict_field",
-                            "in": "query",
-                            "type": "array",
-                            "required": True,
                             "items": {"type": "string"},
                             "collectionFormat": "multi",
                         },
                         {
-                            "name": "list_field",
+                            "name": "date_str",
                             "in": "query",
                             "type": "array",
-                            "required": True,
+                            "format": "date",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "datetime_str",
+                            "in": "query",
+                            "type": "array",
+                            "format": "date-time",
                             "items": {"type": "string"},
                             "collectionFormat": "multi",
                         },
@@ -138,32 +132,25 @@ def test_open_api_definition(client):
                     "operationId": "get_test_resource",
                     "parameters": [
                         {
-                            "name": "float_key",
+                            "name": "key",
                             "in": "query",
                             "type": "array",
-                            "items": {"type": "number"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "float_with_default",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "number"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "dict_field",
-                            "in": "query",
-                            "type": "array",
-                            "required": True,
                             "items": {"type": "string"},
                             "collectionFormat": "multi",
                         },
                         {
-                            "name": "list_field",
+                            "name": "date_str",
                             "in": "query",
                             "type": "array",
-                            "required": True,
+                            "format": "date",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "datetime_str",
+                            "in": "query",
+                            "type": "array",
+                            "format": "date-time",
                             "items": {"type": "string"},
                             "collectionFormat": "multi",
                         },
@@ -198,82 +185,67 @@ def test_open_api_definition(client):
         "tags": [{"name": "Test"}],
         "definitions": {
             "TestCollection_PostRequestModel": {
-                "required": ["dict_field", "list_field"],
                 "properties": {
-                    "dict_field": {
-                        "type": "object",
+                    "date_str": {
+                        "type": "string",
+                        "format": "date",
                         "readOnly": False,
-                        "example": {
-                            "1st dict_field key": "1st dict_field sample",
-                            "2nd dict_field key": "2nd dict_field sample",
-                        },
+                        "example": "2017-09-24",
                     },
-                    "float_key": {"type": "number", "readOnly": False, "example": 1.4},
-                    "float_with_default": {
-                        "type": "number",
+                    "datetime_str": {
+                        "type": "string",
+                        "format": "date-time",
                         "readOnly": False,
-                        "default": 34,
-                        "example": 34,
+                        "example": "2017-09-24T15:36:09",
                     },
-                    "list_field": {
-                        "type": "array",
+                    "key": {
+                        "type": "string",
                         "readOnly": False,
-                        "example": ["1st list_field sample", "2nd list_field sample"],
-                        "items": {"type": "string"},
+                        "example": "sample key",
                     },
                 },
                 "type": "object",
             },
             "TestCollection_PutRequestModel": {
-                "required": ["dict_field", "list_field"],
                 "properties": {
-                    "dict_field": {
-                        "type": "object",
+                    "date_str": {
+                        "type": "string",
+                        "format": "date",
                         "readOnly": False,
-                        "example": {
-                            "1st dict_field key": "1st dict_field sample",
-                            "2nd dict_field key": "2nd dict_field sample",
-                        },
+                        "example": "2017-09-24",
                     },
-                    "float_key": {"type": "number", "readOnly": False, "example": 1.4},
-                    "float_with_default": {
-                        "type": "number",
+                    "datetime_str": {
+                        "type": "string",
+                        "format": "date-time",
                         "readOnly": False,
-                        "default": 34,
-                        "example": 34,
+                        "example": "2017-09-24T15:36:09",
                     },
-                    "list_field": {
-                        "type": "array",
+                    "key": {
+                        "type": "string",
                         "readOnly": False,
-                        "example": ["1st list_field sample", "2nd list_field sample"],
-                        "items": {"type": "string"},
+                        "example": "sample key",
                     },
                 },
                 "type": "object",
             },
             "TestCollection_GetResponseModel": {
-                "required": ["dict_field", "list_field"],
                 "properties": {
-                    "dict_field": {
-                        "type": "object",
+                    "date_str": {
+                        "type": "string",
+                        "format": "date",
                         "readOnly": False,
-                        "example": {
-                            "1st dict_field key": "1st dict_field sample",
-                            "2nd dict_field key": "2nd dict_field sample",
-                        },
+                        "example": "2017-09-24",
                     },
-                    "float_key": {"type": "number", "readOnly": False, "example": 1.4},
-                    "float_with_default": {
-                        "type": "number",
+                    "datetime_str": {
+                        "type": "string",
+                        "format": "date-time",
                         "readOnly": False,
-                        "default": 34,
-                        "example": 34,
+                        "example": "2017-09-24T15:36:09",
                     },
-                    "list_field": {
-                        "type": "array",
+                    "key": {
+                        "type": "string",
                         "readOnly": False,
-                        "example": ["1st list_field sample", "2nd list_field sample"],
-                        "items": {"type": "string"},
+                        "example": "sample key",
                     },
                 },
                 "type": "object",
@@ -284,76 +256,3 @@ def test_open_api_definition(client):
             "MaskError": {"description": "When any error occurs on mask"},
         },
     }
-
-
-def test_post_float_as_int(controller):
-    assert controller.post(
-        {
-            "dict_field": {"any_key": 5},
-            "float_key": 1,
-            "list_field": [22, "33", 44.55, True],
-        }
-    ) == {
-        "dict_field": {"any_key": 5},
-        "float_key": 1,
-        "float_with_default": 34,
-        "list_field": [22, "33", 44.55, True],
-    }
-
-
-def test_get_float_as_int(controller):
-    controller.post(
-        {
-            "dict_field": {"any_key": 5},
-            "float_key": 1,
-            "list_field": [22, "33", 44.55, True],
-        }
-    )
-    assert controller.get_one({"float_key": 1}) == {
-        "dict_field": {"any_key": 5},
-        "float_key": 1,
-        "float_with_default": 34,
-        "list_field": [22, "33", 44.55, True],
-    }
-
-
-def test_put_float_as_int(controller):
-    controller.post(
-        {
-            "dict_field": {"any_key": 5},
-            "float_key": 1,
-            "list_field": [22, "33", 44.55, True],
-        }
-    )
-    assert controller.put(
-        {"dict_field.any_key": 6, "float_key": 1, "float_with_default": 35}
-    ) == (
-        {
-            "dict_field": {"any_key": 5},
-            "float_key": 1,
-            "float_with_default": 34,
-            "list_field": [22, "33", 44.55, True],
-        },
-        {
-            "dict_field": {"any_key": 6},
-            "float_key": 1,
-            "float_with_default": 35,
-            "list_field": [22, "33", 44.55, True],
-        },
-    )
-
-
-def test_get_with_required_field_as_none_is_invalid(controller):
-    controller.post(
-        {
-            "dict_field": {"any_key": 5},
-            "float_key": 1,
-            "list_field": [22, "33", 44.55, True],
-        }
-    )
-    with pytest.raises(layabase.ValidationFailed) as exception_info:
-        controller.get({"dict_field": None})
-    assert exception_info.value.errors == {
-        "dict_field": ["Missing data for required field."]
-    }
-    assert exception_info.value.received_data == {"dict_field": None}
