+++ tests/test_mongo_controller_date.py
@@ -1,9 +1,6 @@
 import datetime
 
-import flask
-import flask_restplus
 import pytest
-from layaberr import ValidationFailed, ModelCouldNotBeFound
 
 import layabase
 import layabase.mongo
@@ -21,37 +18,6 @@ def controller() -> layabase.CRUDController:
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])
     return controller
-
-
-@pytest.fixture
-def app(controller: layabase.CRUDController):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
-        def get(self):
-            return []
-
-        @namespace.expect(controller.json_post_model)
-        def post(self):
-            return []
-
-        @namespace.expect(controller.json_put_model)
-        def put(self):
-            return []
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return []
-
-    return application
 
 
 def test_put_is_updating_date(controller: layabase.CRUDController):
@@ -108,7 +74,7 @@ def test_get_date_is_handled_for_valid_date(controller: layabase.CRUDController)
 
 
 def test_post_invalid_date_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post(
             {
                 "key": "my_key1",
@@ -125,14 +91,14 @@ def test_post_invalid_date_is_invalid(controller: layabase.CRUDController):
 
 
 def test_get_invalid_date_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.get({"date_str": "this is not a date"})
     assert exception_info.value.errors == {"date_str": ["Not a valid date."]}
     assert exception_info.value.received_data == {"date_str": "this is not a date"}
 
 
 def test_delete_invalid_date_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.delete({"date_str": "this is not a date"})
     assert exception_info.value.errors == {"date_str": ["Not a valid date."]}
     assert exception_info.value.received_data == {"date_str": "this is not a date"}
@@ -195,13 +161,13 @@ def test_put_unexisting_is_invalid(controller: layabase.CRUDController):
             "datetime_str": "2016-09-23T23:59:59",
         }
     )
-    with pytest.raises(ModelCouldNotBeFound) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"key": "my_key2"})
     assert exception_info.value.requested_data == {"key": "my_key2"}
 
 
 def test_post_invalid_datetime_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post(
             {
                 "key": "my_key1",
@@ -273,208 +239,3 @@ def test_get_date_is_handled_for_unused_datetime(controller: layabase.CRUDContro
     )
     dt = datetime.datetime.strptime("2016-09-24T23:59:59", "%Y-%m-%dT%H:%M:%S")
     assert controller.get({"datetime_str": dt}) == []
-
-
-def test_open_api_definition(client):
-    response = client.get("/swagger.json")
-    assert response.json == {
-        "swagger": "2.0",
-        "basePath": "/",
-        "paths": {
-            "/test": {
-                "post": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "post_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PostRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "put": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "put_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PutRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_GetResponseModel"
-                            },
-                        }
-                    },
-                    "operationId": "get_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                        {
-                            "name": "X-Fields",
-                            "in": "header",
-                            "type": "string",
-                            "format": "mask",
-                            "description": "An optional fields mask",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            }
-        },
-        "info": {"title": "API", "version": "1.0"},
-        "produces": ["application/json"],
-        "consumes": ["application/json"],
-        "tags": [{"name": "Test"}],
-        "definitions": {
-            "TestCollection_PostRequestModel": {
-                "properties": {
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_PutRequestModel": {
-                "properties": {
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_GetResponseModel": {
-                "properties": {
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                },
-                "type": "object",
-            },
-        },
-        "responses": {
-            "ParseError": {"description": "When a mask can't be parsed"},
-            "MaskError": {"description": "When any error occurs on mask"},
-        },
-    }
