+++ tests/test_mongo_controller_dict.py
@@ -1,7 +1,5 @@
 import enum
 
-import flask
-import flask_restplus
 import pytest
 
 import layabase
@@ -30,28 +28,6 @@ def controller() -> layabase.CRUDController:
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])
     return controller
-
-
-@pytest.fixture
-def app(controller: layabase.CRUDController):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test_parsers")
-    class TestParsersResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        def get(self):
-            return controller.query_get_parser.parse_args()
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return controller.query_delete_parser.parse_args()
-
-    return application
 
 
 def test_post_dict_is_valid(controller: layabase.CRUDController):
@@ -294,27 +270,3 @@ def test_post_dict_is_invalid(controller):
         "key": "my_key",
         "dict_col": {"first_key": "Value1"},
     } == exception_info.value.received_data
-
-
-def test_query_get_parser_with_dict(client):
-    response = client.get(
-        "/test_parsers?dict_col.first_key=2&dict_col.second_key=3&key=4&limit=1&offset=0"
-    )
-    assert response.json == {
-        "dict_col.first_key": ["2"],
-        "dict_col.second_key": [3],
-        "key": ["4"],
-        "limit": 1,
-        "offset": 0,
-    }
-
-
-def test_query_delete_parser_with_dict(client):
-    response = client.delete(
-        "/test_parsers?dict_col.first_key=2&dict_col.second_key=3&key=4"
-    )
-    assert response.json == {
-        "dict_col.first_key": ["2"],
-        "dict_col.second_key": [3],
-        "key": ["4"],
-    }
