+++ tests/test_mongo_controller_dict_without_fields.py
@@ -1,9 +1,6 @@
 import enum
 
-import flask
-import flask_restplus
 import pytest
-from layaberr import ValidationFailed
 
 import layabase
 import layabase.mongo
@@ -30,47 +27,6 @@ def controller():
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])
     return controller
-
-
-@pytest.fixture
-def app(controller):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
-        def get(self):
-            return []
-
-        @namespace.expect(controller.json_post_model)
-        def post(self):
-            return []
-
-        @namespace.expect(controller.json_put_model)
-        def put(self):
-            return []
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return []
-
-    @namespace.route("/test_parsers")
-    class TestParsersResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        def get(self):
-            return controller.query_get_parser.parse_args()
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return controller.query_delete_parser.parse_args()
-
-    return application
 
 
 def test_post_dict_is_valid(controller):
@@ -142,7 +98,7 @@ def test_update_with_dot_notation_invalid_value_is_invalid(controller):
     } == controller.post(
         {"key": "my_key", "dict_col": {"first_key": "Value1", "second_key": 3}}
     )
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"key": "my_key", "dict_col.second_key": "invalid integer"})
     assert {"dict_col.second_key": ["Not a valid int."]} == exception_info.value.errors
     assert {
@@ -158,7 +114,7 @@ def test_delete_with_dot_notation_invalid_value_is_invalid(controller):
     } == controller.post(
         {"key": "my_key", "dict_col": {"first_key": "Value1", "second_key": 3}}
     )
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.delete({"dict_col.second_key": "invalid integer"})
     assert {"dict_col.second_key": ["Not a valid int."]} == exception_info.value.errors
     assert {
@@ -187,7 +143,7 @@ def test_delete_with_dot_notation_enum_value_is_valid(controller):
 
 
 def test_post_with_dot_notation_invalid_value_is_invalid(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post(
             {
                 "key": "my_key",
@@ -273,7 +229,7 @@ def test_put_without_primary_key_is_invalid(controller):
     controller.post(
         {"key": "my_key", "dict_col": {"first_key": "Value1", "second_key": 3}}
     )
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"dict_col": {"first_key": "Value2", "second_key": 4}})
     assert {"key": ["Missing data for required field."]} == exception_info.value.errors
     assert {
@@ -304,7 +260,7 @@ def test_put_dict_with_dot_notation_is_valid(controller):
 
 
 def test_post_dict_is_invalid(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"key": "my_key", "dict_col": {"first_key": "Value1"}})
     assert {
         "dict_col.second_key": ["Missing data for required field."]
@@ -313,200 +269,3 @@ def test_post_dict_is_invalid(controller):
         "key": "my_key",
         "dict_col": {"first_key": "Value1"},
     } == exception_info.value.received_data
-
-
-def test_open_api_definition(client):
-    response = client.get("/swagger.json")
-    assert response.json == {
-        "basePath": "/",
-        "consumes": ["application/json"],
-        "definitions": {
-            "TestCollection_GetResponseModel": {
-                "properties": {
-                    "dict_col": {"default": {}, "readOnly": False, "type": "object"},
-                    "key": {
-                        "example": "sample " "key",
-                        "readOnly": False,
-                        "type": "string",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_PostRequestModel": {
-                "properties": {
-                    "dict_col": {"default": {}, "readOnly": False, "type": "object"},
-                    "key": {
-                        "example": "sample " "key",
-                        "readOnly": False,
-                        "type": "string",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_PutRequestModel": {
-                "properties": {
-                    "dict_col": {"default": {}, "readOnly": False, "type": "object"},
-                    "key": {
-                        "example": "sample " "key",
-                        "readOnly": False,
-                        "type": "string",
-                    },
-                },
-                "type": "object",
-            },
-        },
-        "info": {"title": "API", "version": "1.0"},
-        "paths": {
-            "/test": {
-                "delete": {
-                    "operationId": "delete_test_resource",
-                    "parameters": [
-                        {
-                            "collectionFormat": "multi",
-                            "in": "query",
-                            "items": {"type": "string"},
-                            "name": "key",
-                            "type": "array",
-                        }
-                    ],
-                    "responses": {"200": {"description": "Success"}},
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "operationId": "get_test_resource",
-                    "parameters": [
-                        {
-                            "collectionFormat": "multi",
-                            "in": "query",
-                            "items": {"type": "string"},
-                            "name": "key",
-                            "type": "array",
-                        },
-                        {
-                            "exclusiveMinimum": True,
-                            "in": "query",
-                            "minimum": 0,
-                            "name": "limit",
-                            "type": "integer",
-                        },
-                        {
-                            "in": "query",
-                            "minimum": 0,
-                            "name": "offset",
-                            "type": "integer",
-                        },
-                        {
-                            "description": "An optional " "fields mask",
-                            "format": "mask",
-                            "in": "header",
-                            "name": "X-Fields",
-                            "type": "string",
-                        },
-                    ],
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_GetResponseModel"
-                            },
-                        }
-                    },
-                    "tags": ["Test"],
-                },
-                "post": {
-                    "operationId": "post_test_resource",
-                    "parameters": [
-                        {
-                            "in": "body",
-                            "name": "payload",
-                            "required": True,
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PostRequestModel"
-                            },
-                        }
-                    ],
-                    "responses": {"200": {"description": "Success"}},
-                    "tags": ["Test"],
-                },
-                "put": {
-                    "operationId": "put_test_resource",
-                    "parameters": [
-                        {
-                            "in": "body",
-                            "name": "payload",
-                            "required": True,
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PutRequestModel"
-                            },
-                        }
-                    ],
-                    "responses": {"200": {"description": "Success"}},
-                    "tags": ["Test"],
-                },
-            },
-            "/test_parsers": {
-                "delete": {
-                    "operationId": "delete_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "collectionFormat": "multi",
-                            "in": "query",
-                            "items": {"type": "string"},
-                            "name": "key",
-                            "type": "array",
-                        }
-                    ],
-                    "responses": {"200": {"description": "Success"}},
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "operationId": "get_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "collectionFormat": "multi",
-                            "in": "query",
-                            "items": {"type": "string"},
-                            "name": "key",
-                            "type": "array",
-                        },
-                        {
-                            "exclusiveMinimum": True,
-                            "in": "query",
-                            "minimum": 0,
-                            "name": "limit",
-                            "type": "integer",
-                        },
-                        {
-                            "in": "query",
-                            "minimum": 0,
-                            "name": "offset",
-                            "type": "integer",
-                        },
-                    ],
-                    "responses": {"200": {"description": "Success"}},
-                    "tags": ["Test"],
-                },
-            },
-        },
-        "produces": ["application/json"],
-        "responses": {
-            "MaskError": {"description": "When any error occurs on mask"},
-            "ParseError": {"description": "When a mask can't be parsed"},
-        },
-        "swagger": "2.0",
-        "tags": [{"name": "Test"}],
-    }
-
-
-def test_query_get_parser_with_dict(client):
-    response = client.get(
-        "/test_parsers?dict_col.first_key=2&dict_col.second_key=3&key=4&limit=1&offset=0"
-    )
-    assert response.json == {"key": ["4"], "limit": 1, "offset": 0}
-
-
-def test_query_delete_parser_with_dict(client):
-    response = client.delete(
-        "/test_parsers?dict_col.first_key=2&dict_col.second_key=3&key=4"
-    )
-    assert response.json == {"key": ["4"]}
