+++ tests/test_mongo_controller_int_float.py
@@ -1,5 +1,4 @@
 import pytest
-from layaberr import ValidationFailed
 
 import layabase
 import layabase.mongo
@@ -59,7 +58,7 @@ def test_delete_float_str_in_float_column(controller):
 
 
 def test_post_with_non_int_str_in_int_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"int_value": "abc", "float_value": 1.0})
     assert exception_info.value.errors == {"int_value": ["Not a valid int."]}
     assert exception_info.value.received_data == {
@@ -69,14 +68,14 @@ def test_post_with_non_int_str_in_int_column(controller):
 
 
 def test_post_with_non_float_str_in_float_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"int_value": 1, "float_value": "abc"})
     assert exception_info.value.errors == {"float_value": ["Not a valid float."]}
     assert exception_info.value.received_data == {"float_value": "abc", "int_value": 1}
 
 
 def test_get_with_non_int_str_in_int_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.get({"int_value": "abc", "float_value": 1.0})
     assert exception_info.value.errors == {"int_value": ["Not a valid int."]}
     assert exception_info.value.received_data == {
@@ -86,14 +85,14 @@ def test_get_with_non_int_str_in_int_column(controller):
 
 
 def test_get_with_non_float_str_in_float_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.get({"int_value": 1, "float_value": "abc"})
     assert exception_info.value.errors == {"float_value": ["Not a valid float."]}
     assert exception_info.value.received_data == {"float_value": "abc", "int_value": 1}
 
 
 def test_put_with_non_int_str_in_int_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"int_value": "abc", "float_value": 1.0})
     assert exception_info.value.errors == {"int_value": ["Not a valid int."]}
     assert exception_info.value.received_data == {
@@ -103,14 +102,14 @@ def test_put_with_non_int_str_in_int_column(controller):
 
 
 def test_put_with_non_float_str_in_float_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"int_value": 1, "float_value": "abc"})
     assert exception_info.value.errors == {"float_value": ["Not a valid float."]}
     assert exception_info.value.received_data == {"float_value": "abc", "int_value": 1}
 
 
 def test_delete_with_non_int_str_in_int_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.delete({"int_value": "abc", "float_value": 1.0})
     assert exception_info.value.errors == {"int_value": ["Not a valid int."]}
     assert exception_info.value.received_data == {
@@ -120,7 +119,7 @@ def test_delete_with_non_int_str_in_int_column(controller):
 
 
 def test_delete_with_non_float_str_in_float_column(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.delete({"int_value": 1, "float_value": "abc"})
     assert exception_info.value.errors == {"float_value": ["Not a valid float."]}
     assert exception_info.value.received_data == {"float_value": "abc", "int_value": 1}
