+++ tests/test_mongo_controller_list.py
@@ -1,7 +1,5 @@
 import enum
 
-import flask
-import flask_restplus
 import pytest
 
 import layabase
@@ -14,7 +12,7 @@ class EnumTest(enum.Enum):
 
 
 @pytest.fixture
-def controller():
+def controller() -> layabase.CRUDController:
     class TestCollection:
         __collection_name__ = "test"
 
@@ -32,343 +30,6 @@ def controller():
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])
     return controller
-
-
-@pytest.fixture
-def app(controller):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
-        def get(self):
-            return []
-
-        @namespace.expect(controller.json_post_model)
-        def post(self):
-            return []
-
-        @namespace.expect(controller.json_put_model)
-        def put(self):
-            return []
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return []
-
-    @namespace.route("/test_parsers")
-    class TestParsersResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        def get(self):
-            return controller.query_get_parser.parse_args()
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return controller.query_delete_parser.parse_args()
-
-    return application
-
-
-def test_open_api_definition(client):
-    response = client.get("/swagger.json")
-    assert response.json == {
-        "swagger": "2.0",
-        "basePath": "/",
-        "paths": {
-            "/test": {
-                "post": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "post_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PostRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "put": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "put_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PutRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "list_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "bool_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "boolean"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_GetResponseModel"
-                            },
-                        }
-                    },
-                    "operationId": "get_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "list_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "bool_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "boolean"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                        {
-                            "name": "X-Fields",
-                            "in": "header",
-                            "type": "string",
-                            "format": "mask",
-                            "description": "An optional fields mask",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-            "/test_parsers": {
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "list_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "bool_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "boolean"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "get_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "list_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "bool_field",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "boolean"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-        },
-        "info": {"title": "API", "version": "1.0"},
-        "produces": ["application/json"],
-        "consumes": ["application/json"],
-        "tags": [{"name": "Test"}],
-        "definitions": {
-            "TestCollection_PostRequestModel": {
-                "properties": {
-                    "bool_field": {
-                        "type": "boolean",
-                        "readOnly": False,
-                        "example": True,
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "list_field": {
-                        "type": "array",
-                        "readOnly": False,
-                        "example": [{"first_key": "Value1", "second_key": 1}],
-                        "items": {
-                            "readOnly": False,
-                            "default": {"first_key": None, "second_key": None},
-                            "example": {"first_key": "Value1", "second_key": 1},
-                            "allOf": [{"$ref": "#/definitions/first_key_second_key"}],
-                        },
-                    },
-                },
-                "type": "object",
-            },
-            "first_key_second_key": {
-                "properties": {
-                    "first_key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "Value1",
-                        "enum": ["Value1", "Value2"],
-                    },
-                    "second_key": {"type": "integer", "readOnly": False, "example": 1},
-                },
-                "type": "object",
-            },
-            "TestCollection_PutRequestModel": {
-                "properties": {
-                    "bool_field": {
-                        "type": "boolean",
-                        "readOnly": False,
-                        "example": True,
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "list_field": {
-                        "type": "array",
-                        "readOnly": False,
-                        "example": [{"first_key": "Value1", "second_key": 1}],
-                        "items": {
-                            "readOnly": False,
-                            "default": {"first_key": None, "second_key": None},
-                            "example": {"first_key": "Value1", "second_key": 1},
-                            "allOf": [{"$ref": "#/definitions/first_key_second_key"}],
-                        },
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_GetResponseModel": {
-                "properties": {
-                    "bool_field": {
-                        "type": "boolean",
-                        "readOnly": False,
-                        "example": True,
-                    },
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "list_field": {
-                        "type": "array",
-                        "readOnly": False,
-                        "example": [{"first_key": "Value1", "second_key": 1}],
-                        "items": {
-                            "readOnly": False,
-                            "default": {"first_key": None, "second_key": None},
-                            "example": {"first_key": "Value1", "second_key": 1},
-                            "allOf": [{"$ref": "#/definitions/first_key_second_key"}],
-                        },
-                    },
-                },
-                "type": "object",
-            },
-        },
-        "responses": {
-            "ParseError": {"description": "When a mask can't be parsed"},
-            "MaskError": {"description": "When any error occurs on mask"},
-        },
-    }
 
 
 def test_post_list_of_dict_is_valid(controller):
@@ -627,25 +288,3 @@ def test_put_without_optional_list_of_dict_is_valid(controller):
             ],
         },
     )
-
-
-def test_query_get_parser_with_list_of_dict(client):
-    response = client.get(
-        "/test_parsers?bool_field=true&key=test&list_field=[1,2]&limit=1&offset=0"
-    )
-    assert response.json == {
-        "bool_field": [True],
-        "key": ["test"],
-        "limit": 1,
-        "list_field": [[1, 2]],
-        "offset": 0,
-    }
-
-
-def test_query_delete_parser_with_list_of_dict(client):
-    response = client.delete("/test_parsers?bool_field=true&key=test&list_field=[1,2]")
-    assert response.json == {
-        "bool_field": [True],
-        "key": ["test"],
-        "list_field": [[1, 2]],
-    }
